backend/requirements.txt
fastapi
firebase-admin
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
opencv-python
python-multipart
python-dotenv
shapely
ultralytics
uvicorn
